Some refactoring to move core functionality away from push dialog module
