+++ config.py
@@ -1,0 +1,3 @@
+OFFLINE="offline"
+HYBRID="hybrid"
+ONLINE="online"
