+++ push_dialog.py
@@ -23,16 +23,15 @@
 from __future__ import print_function
 from __future__ import absolute_import
 
-from datetime import datetime
 import os
-import shutil
 
 from PyQt4 import QtGui, QtCore, uic
-from qgis.core import QgsMapLayerRegistry, QgsProject, QgsOfflineEditing, QgsRasterLayer
+from qgis.core import  QgsProject, QgsOfflineEditing
 from PyQt4.QtGui import QDialogButtonBox, QPushButton
 
-from . import data_source_utils as ds
-from .file_utils import fileparts
+from .export_offline import BASE_SAVE_LOCATION, offline_convert, get_layer_ids_to_offline_convert
+from .data_source_utils import *
+from .config import HYBRID
 
 try:
     from .utils.usb import detect_devices, connect_device, push_file, \
@@ -52,7 +51,6 @@ FORM_CLASS, _ = uic.loadUiType(os.path.join(
     os.path.dirname(__file__), 'push_dialog_base.ui'))
 
 
-BASE_SAVE_LOCATION= os.path.expanduser("~")
 
 class PushDialog(QtGui.QDialog, FORM_CLASS):
     def __init__(self, iface, parent=None):
@@ -63,7 +61,7 @@ class PushDialog(QtGui.QDialog, FORM_CLASS):
         self.project = QgsProject.instance()
         self.project_lbl.setText(self.project.title())
         self.push_btn = QPushButton('Push')
-        if self.project_get_remote_layers():
+        if project_get_remote_layers():
             self.push_btn.clicked.connect(self.show_remote_options)
         else:
             self.push_btn.clicked.connect(self.push_project)
@@ -71,10 +69,10 @@ class PushDialog(QtGui.QDialog, FORM_CLASS):
 
         self.devices = None
         self.refresh_devices()
-        self.suggest_offline_wdg.setEnabled(len(self.project_get_always_online_layers())>0)
+        self.suggest_offline_wdg.setEnabled(len(project_get_always_online_layers())>0)
 
     def show_remote_options(self):
-        dlg = RemoteOptionsDialog(self, remote_layers=self.project_get_remote_layers())
+        dlg = RemoteOptionsDialog(self, remote_layers=project_get_remote_layers())
         dlg.exec_()
 
     def refresh_devices(self):
@@ -97,68 +95,19 @@ class PushDialog(QtGui.QDialog, FORM_CLASS):
         progress = float(sent) / total * 100
         self.progress_bar.setValue(progress)
 
-    def get_layer_ids_to_offline_convert(self, remote_layers, remote_save_mode):
-        layer_ids = []
-        if remote_save_mode == RemoteOptionsDialog.OFFLINE: 
-            for layer in remote_layers:
-                layer_ids.append(layer.id())
-
-        for layer in self.project_get_always_offline_layers():
-            layer_ids.append(layer.id())
-        return layer_ids
-
-    def handle_shpfiles(self, dataPath, shpfile_layers):
-        for shpfile_layer in shpfile_layers:
-            shpfile = shpfile_layer.source()
-            parent, fn, ext = fileparts(shpfile)
-            new_file_path = os.path.join(dataPath, fn + ext)
-            # copy surrounding files as well
-            for ext in ds.SHP_EXTENSIONS:
-                file_path = os.path.join(parent, fn + ext)
-                if os.path.exists(file_path):
-                    shutil.copy(file_path, dataPath)
-            ds.change_layer_data_source(shpfile_layer, new_file_path)
-
-
-    def handle_rasters(self, dataPath, raster_layers):
-        for raster_layer in raster_layers:
-            file_path = raster_layer.source()
-            new_file_path = os.path.join(dataPath, os.path.basename(file_path))
-            shutil.copy(file_path, new_file_path)
-            ds.change_layer_data_source(raster_layer, new_file_path)
-
-    def offline_convert(self, vector_layer_ids, raster_layers, shpfile_layers):
-        dt_tag = datetime.now().strftime("%Y%m%d_%H%M%S")
-        existing_filepath = QgsProject.instance().fileName()
-        existing_fn, ext = os.path.splitext(os.path.basename(existing_filepath)) 
-        dataPath = os.path.join(BASE_SAVE_LOCATION, existing_fn+"_"+dt_tag)
-        if not os.path.exists(dataPath):
-            os.mkdir(dataPath)
-        dbPath = "data.sqlite"
-        success = QgsOfflineEditing().convertToOfflineProject(dataPath, dbPath, vector_layer_ids)
-        if not success:
-            raise Exception("Converting to offline project did not succeed")
-        # Now we have a project state which can be saved as offline project
-        # TODO more file-based vectors ??
-        self.handle_rasters(dataPath, raster_layers)
-        self.handle_shpfiles(dataPath, shpfile_layers)
-        # Now we have a project state which can be saved as offline project
-        QgsProject.instance().write(QtCore.QFileInfo(os.path.join(dataPath, existing_fn+"_offline"+ext)))
-        return dataPath
-
 
     def push_project(self, remote_layers=None, remote_save_mode=None):
 
-        can_only_be_online_layers = self.project_get_always_online_layers()
+        can_only_be_online_layers = project_get_always_online_layers()
         if can_only_be_online_layers:
             self.show_warning_about_layers_that_cant_work_offline(can_only_be_online_layers)
 
-        vector_layer_ids = self.get_layer_ids_to_offline_convert(remote_layers, remote_save_mode)
-        shpfile_layers = self.project_get_shp_layers()
-        raster_layers = self.project_get_raster_layers()
-        project_directory = self.offline_convert(vector_layer_ids, raster_layers, shpfile_layers)
+        vector_layer_ids = get_layer_ids_to_offline_convert(remote_layers, remote_save_mode)
+        shpfile_layers = project_get_shp_layers()
+        raster_layers = project_get_raster_layers()
+        project_directory = offline_convert(vector_layer_ids, raster_layers, shpfile_layers)
 
-        if remote_save_mode == RemoteOptionsDialog.HYBRID:
+        if remote_save_mode == HYBRID:
             self.set_hybrid_flag()
         QtGui.QMessageBox.information(self.iface.mainWindow(), 'Info','Please copy {} to your device'.format(project_directory))
         QtGui.QDesktopServices.openUrl(QtCore.QUrl.fromLocalFile(BASE_SAVE_LOCATION))
@@ -176,47 +125,8 @@ class PushDialog(QtGui.QDialog, FORM_CLASS):
         QtGui.QMessageBox.information(self.iface.mainWindow(), 'Warning','Layers {} require a real time connection'.format(layers_list))
 
     def set_hybrid_flag(self):
-        QgsProject.instance().writeEntry(QgsOfflineEditing.PROJECT_ENTRY_SCOPE_OFFLINE,"REMOTE_LAYER_MODE",RemoteOptions.HYBRID)
+        QgsProject.instance().writeEntry(QgsOfflineEditing.PROJECT_ENTRY_SCOPE_OFFLINE,
+                                         "REMOTE_LAYER_MODE",RemoteOptionsDialog.HYBRID)
 
     def on_reload_devices_btn_clicked(self):
         self.refresh_devices()
-
-    @staticmethod
-    def project_filter_layers(filter_func):
-        map_layers = QgsMapLayerRegistry.instance().mapLayers()
-        return [layer for name, layer in map_layers.items() if filter_func(layer)]
-
-
-    @staticmethod
-    def project_get_raster_layers():
-        return PushDialog.project_filter_layers(lambda layer: isinstance(layer, QgsRasterLayer))
-
-    @staticmethod
-    def project_get_shp_layers():
-        return PushDialog.project_filter_layers(ds.is_shapefile_layer)
-
-    @staticmethod
-    def project_get_layers_of_given_types(types):
-        # can see all types via
-        # QgsProviderRegistry.instance().providerList()
-        map_layers = QgsMapLayerRegistry.instance().mapLayers()
-        return [layer for name, layer in map_layers.items() if layer.providerType() in\
-                types and not isinstance(layer, QgsRasterLayer)]
-
-
-    @staticmethod
-    def project_get_always_online_layers():
-        """ Layers that can't be made offline by the offline plugin """
-        online_types = ["WFS", "wcs", "wms", "mssql", "ows"]
-        return PushDialog.project_get_layers_of_given_types(online_types)
-
-    @staticmethod
-    def project_get_remote_layers():
-        """ Remote layers are layers that can either be converted to offline or kept in a realtime or hybrid mode"""
-        return PushDialog.project_get_layers_of_given_types(types=["postgres"])
-
-    @staticmethod
-    def project_get_always_offline_layers():
-        """ Layers that are file based and hence can always be handled by the offline plugin"""
-        types = ['delimitedtext', u'gdal', u'gpx', u'memory', u'ogr', u'spatialite']
-        return PushDialog.project_get_layers_of_given_types(types=types)
